Begin the weakening lemma
For this it was necessary to introduce a new definition with its set of
lemmas, which is the update_tail.
Reading the paper I noticed that somethimes they introduce variables to
the context by the left, and sometimes by the right. For this lemma it is
introduced by the right.

In practice, variables introduced in the right will not shadow variables
already defined in the context. Whereas variables introduced in the left
will.