Changing extds to use app
Then we should check conflicts via wf_extd.
I'm not sure any of this will actually solve the problem though, because
when we are about to use the lemma, we shall now prove wf_extd Gamma xs,
which might be impossible.

Thus I'll stop here and read [Wright and Felleisen 1994] as pointed by the
paper, and try to have a better grasp on this environment issue.