Final context and nonrepeatable fargs
We will use the context such that we must keep its consistence, its the
correct way even looking at good papers.
And we took a look on the used lemmas and it will keep compatibility.

Also within the methodDecl definition we added the hipothesis that the
fargs must not be declared twice. This is important to keep the context
well formed within the term substitution lemma R-INVK case (or something
like that).

This lead to the correct definition of the nonrepeatability of methods
within a classDeclaration. Allowing us to get rid of the very annoying dec
lemmas. (How didn't I see this before? lol)